+++ src/arch/riscv/irqc/aia/aplic.c
@@ -60,7 +60,7 @@ void aplic_init(void)
 
 void aplic_idc_init(void)
 {
-    uint32_t idc_index = cpu()->id;
+    uint32_t idc_index = (uint32_t)(cpu()->id);
     aplic_idc[idc_index].ithreshold = APLIC_IDC_ITHRESHOLD_EN_ALL;
     aplic_idc[idc_index].iforce = APLIC_DISABLE_IFORCE;
     aplic_idc[idc_index].idelivery = APLIC_ENABLE_IDELIVERY;
@@ -144,15 +144,15 @@ void aplic_clr_enbl_reg(size_t reg_indx, uint32_t reg_val)
 
 void aplic_set_target_prio(irqid_t intp_id, uint8_t prio)
 {
-    aplic_control->target[intp_id - 1] &= ~(APLIC_TARGET_IPRIO_MASK);
-    aplic_control->target[intp_id - 1] |= (prio & APLIC_TARGET_IPRIO_MASK);
+    aplic_control->target[intp_id - 1] &= (uint32_t)(~(APLIC_TARGET_IPRIO_MASK));
+    aplic_control->target[intp_id - 1] |= (uint32_t)(prio & APLIC_TARGET_IPRIO_MASK);
 }
 
 void aplic_set_target_hart(irqid_t intp_id, cpuid_t hart)
 {
     aplic_control->target[intp_id - 1] &=
         ~(APLIC_TARGET_HART_IDX_MASK << APLIC_TARGET_HART_IDX_SHIFT);
-    aplic_control->target[intp_id - 1] |= hart << APLIC_TARGET_HART_IDX_SHIFT;
+    aplic_control->target[intp_id - 1] |= (uint32_t)hart << APLIC_TARGET_HART_IDX_SHIFT;
 }
 
 uint8_t aplic_get_target_prio(irqid_t intp_id)
