+++ src/arch/riscv/mem.c
@@ -28,7 +28,7 @@ static inline void as_map_physical_identity(struct addr_space* as)
 
         paddr_t addr = base;
         for (unsigned int j = 0; j < num_entries; j++) {
-            int index = pt_getpteindex_by_va(&as->pt, (vaddr_t)addr, lvl);
+            size_t index = pt_getpteindex_by_va(&as->pt, (vaddr_t)addr, lvl);
             pte_set(&pt[index], addr, PTE_SUPERPAGE, PTE_HYP_FLAGS);
             addr += lvl_size;
         }
@@ -53,7 +53,7 @@ bool mem_translate(struct addr_space* as, vaddr_t va, paddr_t* pa)
             break;
         }
         pte = (pte_t*)pte_addr(pte);
-        int index = pt_getpteindex_by_va(&as->pt, va, i + 1);
+        size_t index = pt_getpteindex_by_va(&as->pt, va, i + 1);
         pte = &pte[index];
     }
     if (pte && pte_valid(pte)) {
