+++ src/platform/drivers/pl011_uart/pl011_uart.c
@@ -96,5 +96,5 @@ void uart_putc(volatile struct Pl011_Uart_hw* ptr_uart, int8_t c)
     // wait until txFIFO is not full
     while (ptr_uart->flag & UART_FR_TXFF) { }
 
-    ptr_uart->data = c;
+    ptr_uart->data = (uint32_t)c;
 }
